Revise assertions and fix bug in the implementation of shifts
Previously failing assertions in the implementations of
ShiftMintOpInstr::EmitNativeCode for various architectures are revised:
* Incorrect assertion !has_shift_count_check() is removed
* Assertion for shift >= 0 is restored
* Assertion for shift < 64 on X64 is moved to kSHL branch

Also, misleading ShiftMintopInstr::has_shift_count_check() method
is replaced with method IsShiftCountInRange() with opposite meaning.

On X64, bug in the implementation of right shifts with constant shift
amount >= 64 is fixed. Corresponding test case is added.

R=regis@google.com, zra@google.com

Review-Url: https://codereview.chromium.org/2974953002 .