Avoid Bigint literals in Dart core library
This CL is a step towards preparation of Dart core library for 64-bit
integers in Dart. In the new --limit-ints-to-64-bits mode integer
literals which are out of range will be rejected, so such literals
should not be used in the core library.

R=zra@google.com

Issue: https://github.com/dart-lang/sdk/issues/30103
Review-Url: https://codereview.chromium.org/2983633002 .