Additional fixes for invalid overrides in dart2js
Together with Jen's fix for checking mixin, these are the remaining cases in
dart2js that we need to silence strong mode errors.

For modelz - the plan is to eventually delete this code, so ignoring makes sense
here. For the namer, Stephen believes we should get rid of the special
subclasses, for now I'm just supressing the error.

R=johhniwinther@google.com, sra@google.com
CC=jmesserly@google.com

Review-Url: https://codereview.chromium.org/2987783002 .