Add dump-partial tool.
This script lets you dump as text a partial .dill file generated by the
incremental generator. To do so, it requires that you pass in other .dill files
that are needed (in order).

In the context of a hot-reload test, you should be able to run this as follows:

pkg/front_end/tool/fasta dump-ir newProgram.dill fullProgram.dill platform.dill

BUG=
R=scheglov@google.com

Review-Url: https://codereview.chromium.org/2985113002 .