Migrate status file entries when migrating files. Yay!
This looks at all the files in the batch being migrated. It goes through
the status files in the 1.0 and strong directories. If it sees an entry
that matches that file, it adds it to the resulting "_2" status file.

It handles adding entries to an existing section if already present in
the new status file. Otherwise, it adds a new section at the end with
the entries.

R=bkonyi@google.com

Review-Url: https://codereview.chromium.org/2989033002 .