Revert status file changes from 58a6227198d2aa927df19f1b4862f8d0017d6da4
R=dmitryas@google.com
TBR=jmesserly@google.com, rnystrom@google.com

The *_strong.status files are used by the analyzer_strong bots
when running on the *_strong suites, so they can't be changed
to reflect ddc status this way.  Separate sections for analyzer
and ddc are needed, if they will be tested on both, and if they differ.

Are these suites going away, and therefore don't need to be tested with
one or both of the products?

BUG=

Review-Url: https://codereview.chromium.org/2988303002 .