A minimal status file formatter and canonicalizer.
For the migration script, I need a (less hacky) way of programmatically
modifying a status file. As a step along that path, I need the ability
to write a status file to disc given a StatusFile object.

With that, it was a small step to write a little "formatter" that turns
a status file into its canonical representation.

It removes unneeded parentheses in expressions, normalizes expectation
case, and normalizes whitespace. Seems useful.

R=whesse@google.com

Review-Url: https://codereview.chromium.org/2988383002 .