Prepare all unlinked units before linking.
This change expects that build mode is always hermetic, i.e. the all
unlinked units that are used for linking are given us as inputs.

We need to separate computing unlinked units from linking, because
linking is synchronous, and access to unlinked units will becore
asynchronous when we switch to using Analysis Driver.

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review-Url: https://codereview.chromium.org/2994753002 .