[kernel] Cache patch-classes.
Via kernel we create PatchClasses to make sure the correct script is set
for functions and fields. Currently we create a new PatchClass each time
we need one, but often an old one could easily be reused.
This CL implements a simple caching-scheme and often uses a cached version.
Because some are requested out-of-order "duplicate" PatchClasses can still
be created though.

BUG=
R=kmillikin@google.com

Review-Url: https://codereview.chromium.org/2967593002 .