[kernel] Make snapshotting work.
This CL enables snapshotting for kernel.
Basically what it does is that it writes out and reads in a few
previously missing entries in the snapshot file (from  Field, Function
and Script).

Debugging from a snapshot made from kernel will not work as snapshots
does not contain sourcecode (it is normally synthesized from the
TokenStream which doesn't exist in kernel).

BUG=
R=kmillikin@google.com

Review-Url: https://codereview.chromium.org/2993613002 .