[fasta] Issue error for large integer literals
This CL adds error reporting for large integer literals into the
new kernel front-end. It is a part of preparation for limiting
integers to 64 bits in Dart. The new error is reported only
if int.parse() fails to parse integer literal, which could happen
in the --limit-ints-to-64-bits mode.

R=ahe@google.com, scheglov@google.com, sigmund@google.com

Issue: https://github.com/dart-lang/sdk/issues/30103
Review-Url: https://codereview.chromium.org/2989073002 .