Remove race condition in Observatory memory page
During the computation of the hover labels in the memory dashboard graph
the format functions were reading shared state that can be out of sync
if an update is ongoing.

To fix the error is enough to use the local copies instead.
To prevent this error in the future the format functions now are marked
as static and the analyzer will signal if they are accessing shared state.

R=asiva@google.com, rmacnak@google.com

Review-Url: https://codereview.chromium.org/3002603002 .