Begin writing tests of the runtime checks needed for Dart 2.0 soundness.
The tests in this CL illustrate the major use cases where runtime
checks are required.  They are not exhaustive.  In a later CL I will
copy over some of analyzer's internal unit tests, which are much more
thorough.

So far these tests only pass when tested via analyzer, since analyzer
currently is the only platform that does the necessary analysis to
figure out which checks are needed.  In follow up CLs I will begin
introducing code into front_end to determine which checks are needed,
and to record this information in the kernel representation.

R=danrubel@google.com, sigmund@google.com

Review-Url: https://codereview.chromium.org/3002893002 .