Clear StringToken.canonicalizer after each parsing.
Otherwise the canonicalizer grows without limit, and if the user
edits files, we constantly parse new strings, so constantly add new
entries into the cananonicalizer. Basically, it is not friendly for
long running processes. I'm not quite sure why we need to makes it
static anyway, if it can be used only for the same file content. Once
we're done with a file, that is it, we don't need its entries.

Anyway, this fixes a leak in Analysis Server.

R=ahe@google.com, brianwilkerson@google.com, danrubel@google.com, paulberry@google.com
BUG=

Review-Url: https://codereview.chromium.org/3003753002 .