[vm, aot] Ensure fixups applied after the world has been compiled are applied to closures in finally clauses.
Finally clauses are inlined into try and catch blocks. This causes the parser to allocate at least two functions for any closure in a finally cause, only one of which will be registered by the flow graph builder. The unregistered function gets compiled, but the program visitor is unable to iterate it. To fix this, we use both the program visitor and the tree-shaker's work queue to apply fixups.

Fixes #30522

R=asiva@google.com, cbernaschina@google.com

Review-Url: https://codereview.chromium.org/3003253002 .