[gardening] Re-enable vm/cc/UseDartApi on simdbc64
Running this in a loop locally did not reproduce any crash and the issue
was filed in January, so the fundamental problem might have been fixed
already.  Due to the lack of more information in the bug report, the
best thing is to renable the test, close the issue and see if the test
flakes again.

Closes https://github.com/dart-lang/sdk/issues/28499

R=vegorov@google.com

Review-Url: https://codereview.chromium.org/3003203002 .