Remove check for scheme in source-loader.
Context: this prevents us from using other schemes in unit tests, and causes
issues when processign a URI with no scheme, that happens to be supported by the
underlying file-system (e.g. PhysicalFileSystem will make it a `file:*` URI if
no scheme was present).

This is necessary to fix dartbug.com/30141

Review-Url: https://codereview.chromium.org/3006483002 .