Adding adjusted `void` tests from CL 2699073003
These tests exercise the static and dynamic semantics of `void` used in
those new locations where 'generalized void' added support form it.

They came from CL 2699073003 and were adjusted to match the result of
discussions taken after their creation (in particular, `b ? e1 : e2`
was used and expected to be allowed, but it is now a static warning in
Dart 1.x and a compile-time error in Dart 2).

The tests are located in tests/language_2, and they are testing Dart 2
behavior. In particular, usage of a value of static type `void` is a
compile-time error rather than a static warning, and `void` is
considered equivalent to `Object` in subtype checks, both statically
and dynamically.

We do not currently have any tools which will conform to the
expectations of these tests. Status files have been adjested
accordingly.

R=floitsch@google.com

Review-Url: https://codereview.chromium.org/2997353002 .