[VM] Make us compile with debug-optimization-level=0 again
After [0] landed our debug builders managed to still compile, because
the default debug-optimization-level allows the compiler to inline. This
caused a "CanReload() const { return false; }" to be inlined and
constant folded, so the call to ReloadSources() was eliminated.

Compiling with debug-optimization-level=0 disables inlining and keeps
the call, which gives us a linker error when building the
dart_precompiled_runtime target.

[0] https://codereview.chromium.org/2997243002

R=vegorov@google.com

Review-Url: https://codereview.chromium.org/3003383002 .